reworked oauth server factoryand handler
